+++ Tests/Get-RubrikAPIData.Tests.ps1
@@ -19,7 +19,7 @@ Describe -Name 'Private/Get-RubrikAPIData' -Tag 'Private', 'Get-RubrikAPIData' -
     }
 
     #Function List
-    $ignorelist = @('Invoke-RubrikRESTCall','Move-RubrikMountVMDK','Sync-RubrikAnnotation','Sync-RubrikTag') 
+    $ignorelist = @('Invoke-RubrikRESTCall','Move-RubrikMountVMDK','Sync-RubrikAnnotation','Sync-RubrikTag','Get-RubrikModuleDefaultParameter','Set-RubrikModuleDefaultParameter','Get-RubrikModuleOption','Set-RubrikModuleOption','Remove-RubrikModuleDefaultParameter')
     $functions = ( Get-ChildItem -Path './Rubrik/Public' | Where-Object extension -eq '.ps1').Name.Replace('.ps1','')
     $functions = $functions | Where-Object {$ignorelist -notcontains $_}
 
@@ -32,18 +32,18 @@ Describe -Name 'Private/Get-RubrikAPIData' -Tag 'Private', 'Get-RubrikAPIData' -
     #endregion
 
     Context -Name "Function specific tests" -Fixture {
-        # This variable contains an array of functions that should be excluded from API endpoint testing 
+        # This variable contains an array of functions that should be excluded from API endpoint testing
         $ExcludedFunctions = @(
             'Get-RubrikObject'
             'Get-RubrikDatabaseRecoveryPoint'
         )
-        
-        it -Name "Get-RubrikAPIData - <f> test" -TestCases $cases { 
+
+        it -Name "Get-RubrikAPIData - <f> test" -TestCases $cases {
             param($v,$f)
             if ($ExcludedFunctions -contains $f) {
                 $methodresult = $true
                 $uriresult = $true
-            } else {            
+            } else {
                 $v | ForEach-Object -Begin {
                     $methodresult = New-Object System.Collections.Generic.List[System.Boolean]
                     $uriresult = New-Object System.Collections.Generic.List[System.Boolean]
@@ -61,12 +61,12 @@ Describe -Name 'Private/Get-RubrikAPIData' -Tag 'Private', 'Get-RubrikAPIData' -
             $uriresult | Should -Contain $true
         }
     }
-    
+
     Context -Name "Should contain Function property as output" {
         It -Name 'Verify property exists' -Test {
             $functions = ( Get-ChildItem -Path './Rubrik/Public' |
                 Where-Object extension -eq '.ps1').Name.Replace('.ps1','')
-            $ignorelist = @('Invoke-RubrikRESTCall','Move-RubrikMountVMDK','Sync-RubrikAnnotation','Sync-RubrikTag','Get-RubrikObject','Get-RubrikDatabaseRecoveryPoint') 
+            $ignorelist = @('Invoke-RubrikRESTCall','Move-RubrikMountVMDK','Sync-RubrikAnnotation','Sync-RubrikTag','Get-RubrikObject','Get-RubrikDatabaseRecoveryPoint','Get-RubrikModuleDefaultParameter','Set-RubrikModuleDefaultParameter','Get-RubrikModuleOption','Set-RubrikModuleOption','Remove-RubrikModuleDefaultParameter')
             $functions = $functions | Where-Object {$ignorelist -notcontains $_}
             $functions | ForEach-Object {
                 (Get-RubrikAPIData -Endpoint $_).Function |